requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[asyncio]==5.2.1
anthropic==0.42.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pydantic-settings==2.7.0
pydantic==2.10.3
httpx==0.28.1
PyPDF2==3.0.1
python-docx==1.1.2
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.155.0
aiofiles==24.1.0
python-dotenv==1.0.1
cryptography==43.0.3
